+++ includes/class-wpsms-gateway.php
@@ -873,6 +873,7 @@ It might be a phone number (e.g., +1 555 123 4567) or an alphanumeric ID if supp
                 'arkapayamak'   => 'arkapayamak.com',
                 'smsservice'    => 'smsservice.ir',
                 'firstpayamak'  => 'firstpayamak.com',
+                'ghasedak'      => 'ghasedak.me',
             ),
             'arabic'               => array(
                 'msegat'       => 'msegat.com',
@@ -1561,7 +1562,7 @@ It might be a phone number (e.g., +1 555 123 4567) or an alphanumeric ID if supp
             $settingsLink = admin_url('admin.php?page=wp-sms-settings&tab=gateway');
 
             $message = sprintf(
-                /* translators: %1$s: URL to gateway settings page */
+            /* translators: %1$s: URL to gateway settings page */
                 __('SMS gateway setup requires your attention. <a href="%1$s">Review and update your gateway settings</a> to ensure SMS messages are sent successfully.', 'wp-sms'),
                 esc_url($settingsLink)
             );